FastAPI/3일차 과제/pyproject.toml
[tool.poetry]
name = "3일차 과제"
version = "0.1.0"
description = ""
authors = ["choiguen <daindj0@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.118.0"
uvicorn = "^0.37.0"
sqlalchemy = "^2.0.43"
requests = "^2.32.5"
pyjwt = "^2.10.1"
jinja2 = "^3.1.6"
python-multipart = "^0.0.20"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
ruff = "^0.13.2"
mypy = "^1.18.2"
pytest = "^8.4.2"
coverage = "^7.10.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
target-version = "py313"

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.13
strict = true

[tool.coverage.run]
omit = ["/test_.py"]